requirements.txt
scrapegraphai
playwright
python-dotenv
groq
langchain-groq 